Add logic for local containers discovery using /proc fs
Discovering containers, running on the local node is essential
functionality for trace-kube. There are two approaches that can be used:
extracting information for namespaces from /proc file system or using
the CRI interface. Both will be supported by trace-kube.
Proposed changes implement the first approach. The /proc file system is
traversed and all processes that run in unique 'uts' and 'pid' name
spaces are considered to be a unique container. All discovered
containers are recorded, together with a list of processes that run in
the container. A new API GET endpoint is added - "/containers", to
expose the discovered containers. That is used mostly for debug purposes
for now, but it could be useful as part of the official API also.

Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>