Manager for trace hooks
Tracing functionality is provided by standalone helper applications,
located in a given directory. In order to use them, trace-kube should
be aware of all available trace hooks and their parameters. At startup
these applications are auto discovered and stored in internal database.
A new GET API is introduced, to get known trace hooks - mostly for
debugging purposes, but it could be part of the official API.

Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>