tracer: Store path to /proc fs into pods database
The path to /proc fs is used from pod_proc.go to discover running local
containers using the information from /proc. Instead of hard coding the
path, store it in the pods database and pass it to pod_proc.go at init
time. This allows to use non standard path, which is useful when host
/proc is mounted inside a container.

Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>