tracer: Pass custom /sys mount point
When tracer runs in a container, the host /proc and /sys directories are
mounted on a custom mount point inside the container. These directories
are mandatory for trace hooks. Pass the location of these mount points
to the trace-kube application, so that they can be forwarded to the
trace-hooks. Those can be passed as trace-kube input parameters or as
environment variables.

Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>