Verify CRI endpoint
When CRI endpoint is auto discovered by the tracer application, a list
of well known CRI endpoints is tested for connection. There could be a
case when more than one active CRI endpoints exists on the node. That's
why a logic for CRI endpoint verification is implemented: check if
current tracer pod exists in the CRI database and accept only that
endpoint as valid.

Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>