Do not run the tracers in the host network name space
Back-end tracer pods expose a well-known port 8080 for communication
with the front-end trace-kube service. If the tracers run in the host
network name space, there could be a conflict with some other service,
using the same port. In order to avoid any possible conflict, move the
tracers in its own network name space.

Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>