Directories restructuring
The current codebase is located into the "server" directory. This layout
was designed according to an old idea of the project structure, which is
not relevant any more. Currently, both tracer and service applications
share common directory layout, according to golang recommendations.
That's why the code is moved one level up, in the top project directory.

Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>