Add TLS installation using self-signed certificate
Using secure API is essential for container-tracer, as it runs with root
privileges on the system. When installing on Kubernetes cluster, the
cert-manager component can be used to manage a TLS certificate, required
to run the API on https. Added such installation, using a self-signed
certificate created install time by cert-manager.

Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>